xpu: move host register ops into their own file/tests
Move hostRegister/hostUnregister and the xpu_host_register/
xpu_host_unregister op wrappers out of mem_cpy.cpp into a dedicated
csrc/utils/host_register.cpp, registered in CMakeLists.txt. This keeps
mem_cpy.cpp focused on the memcpy/staging helpers.

Move the corresponding tests out of tests/test_cache.py into a
dedicated tests/test_host_register.py, matching the new source file
layout.
